requirements.txt
torchvision==0.7.0
tqdm==4.48.2
numpy==1.19.1
torch==1.6.0
filelock==3.0.12
Pillow==7.2.0
